+++ pandas_ml/compat.py
@@ -7,9 +7,18 @@ from pandas.api.types import is_list_like, is_integer_dtype         # noqa
 
 PANDAS_VERSION = LooseVersion(pd.__version__)
 
+if PANDAS_VERSION >= LooseVersion('0.23'):
+    _PANDAS_ge_023 = True
+else:
+    _PANDAS_ge_023 = False
+
+if PANDAS_VERSION >= LooseVersion('0.22'):
+    _PANDAS_ge_022 = True
+else:
+    _PANDAS_ge_022 = False
+
 if PANDAS_VERSION >= LooseVersion('0.21'):
     _PANDAS_ge_021 = True
-
 else:
     _PANDAS_ge_021 = False