+++ pandas_ml/misc/patsy_wraps.py
@@ -3,9 +3,12 @@
 
 def transform_with_patsy(formula, data, *args, **kwargs):
     try:
+        # needs patsy v0.5.1 to support formula in Python 3.7
+        # https://github.com/pydata/patsy/pull/131
         import patsy
     except ImportError:
         raise ImportError("'patsy' is required to transform with string formula")
+
     if '~' in formula:
         y, X = patsy.dmatrices(formula, data=data, return_type='dataframe',
                                *args, **kwargs)