+++ pandas_ml/smaccessors/test/test_base.py
@@ -119,11 +119,11 @@ class TestBaseRegressor(tm.TestCase):
             self.assert_numpy_array_almost_equal(result, expected)
 
     def test_gridsearch(self):
-        import sklearn.grid_search as gs
+        import sklearn.model_selection as ms
         tuned_parameters = {'statsmodel': [sm.OLS, sm.GLS]}
         diabetes = datasets.load_diabetes()
 
-        cv = gs.GridSearchCV(base.StatsModelsRegressor(sm.OLS), tuned_parameters, cv=5, scoring=None)
+        cv = ms.GridSearchCV(base.StatsModelsRegressor(sm.OLS), tuned_parameters, cv=5, scoring=None)
         fitted = cv.fit(diabetes.data, diabetes.target)
         self.assertTrue(fitted.best_estimator_.statsmodel is sm.OLS)